fix(api): api 분리
