--- src/react-pandamarket/Context/SearchContext.js
@@ -1,4 +1,0 @@
-import React from "react";
-const SearchContext = React.createContext();
-
-export default SearchContext;
