--- src/react-pandamarket/api/api.js
@@ -1,47 +1,0 @@
-import axios from "axios";
-
-const Base_URL = "https://panda-market-api.vercel.app";
-
-export async function getProduct(
-  page = 1,
-  pageSize = 10,
-  orderBy = "recent",
-  keyword = ""
-) {
-  const query = `page=${page}&pageSize=${pageSize}&orderBy=${orderBy}&keyword=${keyword}`;
-  try {
-    const response = await axios.get(`${Base_URL}/products?${query}`);
-    return response.data;
-  } catch (error) {
-    if (error.response) {
-      throw new Error("데이터를 불러오는데 실패했습니다");
-    }
-  }
-}
-
-export async function getProductDetail(productId = "") {
-  try {
-    const response = await axios.get(`${Base_URL}/products/${productId}`);
-    return response.data;
-  } catch (error) {
-    if (error.response) {
-      throw new Error("데이터를 불러오는데 실패했습니다");
-    }
-  }
-}
-
-export async function getComment(productId, limit = 10, cursor) {
-  console.log(productId);
-  const query = `limit=${limit}&cursor=${cursor}`;
-  try {
-    const response = await axios.get(
-      `${Base_URL}/products/${productId}/comments?${query}`
-    );
-    return response.data;
-  } catch (error) {
-    if (error.response) {
-      console.log(error.message);
-      throw new Error("데이터를 불러오는데 실패했습니다");
-    }
-  }
-}
