+++ src/react-pandamarket/utils/productDetailFunctions.js
@@ -1,20 +1,17 @@
-import { getComment, getProductDetail } from "../api/api";
+import { getProductDetail } from "../api/getProductdetail";
+import { getComment } from "../api/getComment";
 
 // 상품 세부 정보 가져오기
-export async function getDetail( productId = "") {
+export async function getDetail(productId = "") {
   const response = await getProductDetail(productId);
   return response;
 }
 
 // 문의 댓글 가져오기
-export async function comment(
-  productId,
-  limit,
-  cursor,
-) {
+export async function comment(productId, limit, cursor) {
   const response = await getComment(productId, limit, cursor);
   return {
     nextCursor: response.nextCursor,
-    comments: response.list
-  }
+    comments: response.list,
+  };
 }
